requirements.txt
fastapi
uvicorn
streamlit
langchain
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
